Now, there needs to be plots made all data is organized in a population dictionary.
The function can be used for all other stages of the experiment.
So now we can have plots with before's and afters.

Once i figure out how to give matplotlib a dictionary with the key being the dataframe "time" and it the left tuple being a rabbit and the right tuple is a fox